pyproject.toml
[project]
name = "eaupourtous"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "crispy-tailwind>=1.0.3",
    "django>=6.0.3",
    "django-crispy-forms>=2.6",
    "django-modeltranslation>=0.20.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.4.0",
    "weasyprint>=68.1",
]
